Reduce memory usage
git-svn-id: svn+ssh://solsvn.oma.be/home/sidc/ssr/SWHV_repos/trunk/client/jhv-3d-wcs@5013 b4e469a2-07ce-4b26-9273-4d7d95a670c7

+++ src/plugins/jhv-eveplugin/src/org/helioviewer/jhv/plugins/eveplugin/lines/data/DownloadController.java
@@ -324,13 +324,13 @@ public class DownloadController {
                 }
 
                 // Log.warn(data.toString());
-                final double[] values = new double[length];
+                final float[] values = new float[length];
                 final long[] dates = new long[length];
 
                 for (int i = 0; i < length; i++) {
                     final JSONArray entry = data.getJSONArray(i);
                     final long millis = ((long) entry.getDouble(0)) * 1000;
-                    values[i] = entry.getDouble(1) * multiplier;
+                    values[i] = (float) (entry.getDouble(1) * multiplier);
                     dates[i] = millis;
                 }
