+++ src/plugins/jhv-eveplugin/src/org/helioviewer/jhv/plugins/eveplugin/lines/data/EVECacheController.java
@@ -64,7 +64,7 @@ public class EVECacheController {
         controllerListeners.remove(listener);
     }
 
-    public void addToCache(Band band, double[] values, long[] dates) {
+    public void addToCache(Band band, float[] values, long[] dates) {
         cache.add(band, values, dates);
         fireDataAdded(band);
     }
